app: split out root navigation destinations
